Feature/part1

+++ service-module/src/main/java/com/gmail/portnova/julia/service/impl/UserServiceImpl.java
@@ -7,28 +7,27 @@ import com.gmail.portnova.julia.repository.model.RoleNameEnum;
 import com.gmail.portnova.julia.repository.model.User;
 import com.gmail.portnova.julia.service.UserService;
 import com.gmail.portnova.julia.service.converter.GeneralConverter;
+import com.gmail.portnova.julia.service.exception.UserNotFoundException;
+import com.gmail.portnova.julia.service.exception.UserRoleNotFoundException;
+import com.gmail.portnova.julia.service.model.PageDTO;
+import com.gmail.portnova.julia.service.model.PageableUser;
 import com.gmail.portnova.julia.service.model.UserDTO;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
+@RequiredArgsConstructor
 @Service
 public class UserServiceImpl implements UserService {
     private final UserRepository userRepository;
     private final GeneralConverter<User, UserDTO> userConverter;
     private final RoleRepository roleRepository;
-
-
-    public UserServiceImpl(UserRepository userRepository,
-                           GeneralConverter<User, UserDTO> userConverter, RoleRepository roleRepository) {
-        this.userRepository = userRepository;
-        this.userConverter = userConverter;
-        this.roleRepository = roleRepository;
-    }
 
     @Override
     @Transactional
@@ -43,27 +42,15 @@ public class UserServiceImpl implements UserService {
 
     @Override
     @Transactional
-    public List<UserDTO> findAllUsersExceptCurrent(String email, Integer page, Integer maxResult) {
-        int startPosition = maxResult * (page - 1);
-        List<User> users = userRepository.findAllExceptCurrent(email, startPosition, maxResult);
-        return users.stream()
-                .map(userConverter::convertObjectToDTO)
-                .collect(Collectors.toList());
-    }
-
-    @Override
-    @Transactional
-    public Long countPages(Integer maxResult) {
-        Long numberOfRows = userRepository.count();
-        return (numberOfRows / maxResult) + 1;
-    }
-
-    @Override
-    @Transactional
-    public void deleteByUUID(String id) {
+    public UserDTO deleteByUUID(String id) {
         UUID uuid = UUID.fromString(id);
         User user = userRepository.findByUuid(uuid);
-        userRepository.remove(user);
+        if (Objects.nonNull(user)) {
+            userRepository.remove(user);
+            return userConverter.convertObjectToDTO(user);
+        } else {
+            throw new UserNotFoundException(String.format("User with uuid %s was not found", id));
+        }
     }
 
     @Override
@@ -71,7 +58,11 @@ public class UserServiceImpl implements UserService {
     public UserDTO findByUuid(String id) {
         UUID uuid = UUID.fromString(id);
         User user = userRepository.findByUuid(uuid);
-        return userConverter.convertObjectToDTO(user);
+        if (Objects.nonNull(user)) {
+            return userConverter.convertObjectToDTO(user);
+        } else {
+            throw new UserNotFoundException(String.format("User with uuid %s was not found", id));
+        }
     }
 
     @Override
@@ -82,10 +73,48 @@ public class UserServiceImpl implements UserService {
         if (Objects.nonNull(user)) {
             RoleNameEnum roleName = RoleNameEnum.valueOf(newRole);
             Role role = roleRepository.findByName(roleName);
-            user.setRole(role);
-            return userConverter.convertObjectToDTO(user);
+            if (Objects.nonNull(role)) {
+                user.setRole(role);
+                return userConverter.convertObjectToDTO(user);
+            } else {
+                throw new UserRoleNotFoundException(String.format("Role with name %s was not found", newRole));
+            }
         } else {
-            return null;
+            throw new UserNotFoundException(String.format("User with uuid %s was not found", id));
+        }
+    }
+
+    @Override
+    public PageDTO<UserDTO> getUsersPage(String email, Integer pageNumber, Integer maxResult) {
+        Long numberOfRows = userRepository.count();
+        Long numberOfRowsExceptCurrentUser = numberOfRows - 1;
+        PageableUser page = new PageableUser();
+        page.setTotalPages(getNumberOfPages(numberOfRowsExceptCurrentUser, maxResult));
+        int startPosition = getStartPosition(pageNumber, maxResult);
+        List<User> users = userRepository.findAllExceptCurrent(email, startPosition, maxResult);
+        setPageDTOList(page, users);
+        return page;
+    }
+
+    private int getStartPosition(Integer page, Integer maxResult) {
+        return maxResult * (page - 1);
+    }
+
+    private void setPageDTOList(PageableUser page, List<User> users) {
+        List<UserDTO> userDTOS = new ArrayList<>();
+        if (!users.isEmpty()) {
+            userDTOS.addAll(users.stream()
+                    .map(userConverter::convertObjectToDTO)
+                    .collect(Collectors.toList()));
+        }
+        page.getObjects().addAll(userDTOS);
+    }
+
+    private Long getNumberOfPages(Long numberOfRows, Integer maxResult) {
+        if (numberOfRows % maxResult == 0) {
+            return numberOfRows / maxResult;
+        } else {
+            return numberOfRows / maxResult + 1;
         }
     }
 }
